# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BPMS API",
        "description": "API documentation for BPMS app",
        "version": "1.0.0"
    },
    "paths": {
        "/api/cost-center": {
            "get": {
                "tags": [
                    "Cost Center Hierarchy"
                ],
                "summary": "Get All Data From Table MapCostCenterHierarchy",
                "description": "Retrieve all data from the MapCostCenterHierarchy table, including relationships.",
                "operationId": "getAllCostCenterHierarchy",
                "parameters": [
                    {
                        "name": "cost_center",
                        "in": "query",
                        "description": "Filter by cost center ID",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "TCB1000000"
                        }
                    },
                    {
                        "name": "cost_center_name",
                        "in": "query",
                        "description": "Filter by cost center name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "SVP MARKETING"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "cost_center": {
                                                "type": "string",
                                                "example": "AB00000000"
                                            },
                                            "cost_center_name": {
                                                "type": "string",
                                                "example": "Sales Department",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cost Center not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}